:recycle: (916): extract preprocess_audio and post_process_segments into use-case _shared
